+++ src/test/java/org/stellar/sdk/Sep10ChallengeTest.java
@@ -25,12 +25,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         Network.TESTNET,
         client.getAccountId(),
-        "angkor wat",
+        domainName,
         timeBounds
     );
 
@@ -44,7 +45,7 @@ public class Sep10ChallengeTest {
     assertEquals(1, transaction.getOperations().length);
     ManageDataOperation op = (ManageDataOperation) transaction.getOperations()[0];
     assertEquals(client.getAccountId(), op.getSourceAccount());
-    assertEquals("angkor wat auth", op.getName());
+    assertEquals(domainName + " auth", op.getName());
 
     assertEquals(64, op.getValue().length);
     BaseEncoding base64Encoding = BaseEncoding.base64();
@@ -64,12 +65,13 @@ public class Sep10ChallengeTest {
       long now = System.currentTimeMillis() / 1000L;
       long end = now + 300;
       TimeBounds timeBounds = new TimeBounds(now, end);
+      String domainName = "example.com";
 
       Sep10Challenge.newChallenge(
           server,
           Network.TESTNET,
           "MCAAAAAAAAAAAAB7BQ2L7E5NBWMXDUCMZSIPOBKRDSBYVLMXGSSKF6YNPIB7Y77ITKNOG",
-          "angkor wat",
+          domainName,
           timeBounds
       );
       fail();
@@ -84,6 +86,7 @@ public class Sep10ChallengeTest {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
     Network network = Network.TESTNET;
+    String domainName = "example.com";
 
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
@@ -93,11 +96,11 @@ public class Sep10ChallengeTest {
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
-    Sep10Challenge.ChallengeTransaction challengeTransaction = Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET);
+    Sep10Challenge.ChallengeTransaction challengeTransaction = Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET, domainName);
     assertEquals(new Sep10Challenge.ChallengeTransaction(transaction, client.getAccountId()), challengeTransaction);
   }
 
@@ -106,6 +109,7 @@ public class Sep10ChallengeTest {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
     Network network = Network.TESTNET;
+    String domainName = "example.com";
 
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
@@ -115,7 +119,7 @@ public class Sep10ChallengeTest {
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -126,7 +130,8 @@ public class Sep10ChallengeTest {
     Sep10Challenge.ChallengeTransaction v0ChallengeTransaction = Sep10Challenge.readChallengeTransaction(
         v0Base64,
         server.getAccountId(),
-        Network.TESTNET
+        Network.TESTNET,
+        domainName
     );
 
     transaction.setEnvelopeType(EnvelopeType.ENVELOPE_TYPE_TX);
@@ -136,10 +141,20 @@ public class Sep10ChallengeTest {
     Sep10Challenge.ChallengeTransaction v1ChallengeTransaction = Sep10Challenge.readChallengeTransaction(
         v1Base64,
         server.getAccountId(),
-        Network.TESTNET
+        Network.TESTNET,
+        domainName
     );
 
     assertEquals(v0ChallengeTransaction, v1ChallengeTransaction);
+
+    for (String envelopeBase64 : ImmutableList.of(v0Base64, v1Base64)) {
+      Sep10Challenge.ChallengeTransaction challengeTransaction = Sep10Challenge.readChallengeTransaction(
+          envelopeBase64,
+          server.getAccountId(),
+          Network.TESTNET, domainName
+      );
+      assertEquals(new Sep10Challenge.ChallengeTransaction(transaction, client.getAccountId()), challengeTransaction);
+    }
   }
 
   @Test
@@ -151,12 +166,14 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
+
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -164,7 +181,7 @@ public class Sep10ChallengeTest {
       Sep10Challenge.readChallengeTransaction(
           transaction.toEnvelopeXdrBase64(),
           "MCAAAAAAAAAAAAB7BQ2L7E5NBWMXDUCMZSIPOBKRDSBYVLMXGSSKF6YNPIB7Y77ITKNOG",
-          Network.TESTNET
+          Network.TESTNET, domainName
       );
       fail();
     } catch (FormatException e) {
@@ -177,6 +194,7 @@ public class Sep10ChallengeTest {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
     Network network = Network.TESTNET;
+    String domainName = "example.com";
 
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
@@ -186,7 +204,7 @@ public class Sep10ChallengeTest {
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
     Operation[] operations = transaction.getOperations();
@@ -206,7 +224,7 @@ public class Sep10ChallengeTest {
       Sep10Challenge.readChallengeTransaction(
           withMuxedClient.toEnvelopeXdrBase64(),
           "MCAAAAAAAAAAAAB7BQ2L7E5NBWMXDUCMZSIPOBKRDSBYVLMXGSSKF6YNPIB7Y77ITKNOG",
-          Network.TESTNET
+          Network.TESTNET, domainName
       );
       fail();
     } catch (IllegalArgumentException e) {
@@ -223,18 +241,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
+
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(client);
 
-    Sep10Challenge.ChallengeTransaction challengeTransaction = Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET);
+    Sep10Challenge.ChallengeTransaction challengeTransaction = Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET, domainName);
     assertEquals(new Sep10Challenge.ChallengeTransaction(transaction, client.getAccountId()), challengeTransaction);
   }
 
@@ -242,7 +262,7 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidNotSignedByServer() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -256,7 +276,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -273,7 +293,7 @@ public class Sep10ChallengeTest {
 
     transaction.sign(client);
     try {
-      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network);
+      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals(String.format("Transaction not signed by server: %s.", server.getAccountId()), e.getMessage());
@@ -289,19 +309,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     String challenge = transaction.toEnvelopeXdrBase64().replace("A", "B");
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (RuntimeException ignored) {
       // TODO: I think we should add an exception signature about the failure to read XDR on Transaction.fromEnvelopeXdr.
@@ -316,18 +337,19 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         Network.TESTNET,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     // assertThrows requires Java 8+
     try {
-      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), KeyPair.random().getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), KeyPair.random().getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction source account is not equal to server's account.", e.getMessage());
@@ -338,7 +360,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidSeqNoNotZero() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -352,7 +375,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), 100L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -370,7 +393,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("The transaction sequence number should be zero.", e.getMessage());
@@ -381,7 +404,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidTimeboundsInfinite() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -395,7 +419,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation operation = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation operation = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
     Operation[] operations = new Operation[]{operation};
@@ -412,7 +436,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction requires non-infinite timebounds.", e.getMessage());
@@ -423,7 +447,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidNoTimeBounds() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     byte[] nonce = new byte[48];
@@ -433,7 +458,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation operation = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation operation = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
     Operation[] operations = new Operation[]{operation};
@@ -450,7 +475,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction requires timebounds.", e.getMessage());
@@ -466,17 +491,18 @@ public class Sep10ChallengeTest {
     long start = current + 300;
     long end = current + 600;
     TimeBounds timeBounds = new TimeBounds(start, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         Network.TESTNET,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     try {
-      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction is not within range of the specified timebounds.", e.getMessage());
@@ -492,17 +518,18 @@ public class Sep10ChallengeTest {
     long start = current - 600;
     long end = current - 300;
     TimeBounds timeBounds = new TimeBounds(start, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         Network.TESTNET,
         client.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     try {
-      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction is not within range of the specified timebounds.", e.getMessage());
@@ -513,7 +540,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidTooManyOperations() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -527,11 +555,11 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
-    ManageDataOperation manageDataOperation2 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation2 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -549,7 +577,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction requires a single ManageData operation.", e.getMessage());
@@ -565,6 +593,7 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
     SetOptionsOperation setOptionsOperation = new SetOptionsOperation.Builder()
@@ -585,7 +614,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Operation type should be ManageData.", e.getMessage());
@@ -595,7 +624,8 @@ public class Sep10ChallengeTest {
   @Test
   public void testReadChallengeTransactionInvalidOperationNoSourceAccount() throws IOException {
     KeyPair server = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -609,7 +639,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .build();
 
     Operation[] operations = new Operation[]{manageDataOperation1};
@@ -626,7 +656,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Operation should have a source account.", e.getMessage());
@@ -637,7 +667,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidDataValueWrongEncodedLength() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -651,7 +682,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -669,7 +700,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Random nonce encoded as base64 should be 64 bytes long.", e.getMessage());
@@ -680,7 +711,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidDataValueCorruptBase64() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -689,7 +721,7 @@ public class Sep10ChallengeTest {
 
     byte[] encodedNonce = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA?AAAAAAAAAAAAAAAAAAAAAAAAAA".getBytes("UTF-8");
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -707,7 +739,7 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Failed to decode random nonce provided in ManageData operation.", e.getMessage());
@@ -719,7 +751,8 @@ public class Sep10ChallengeTest {
   public void testReadChallengeTransactionInvalidDataValueWrongByteLength() throws IOException {
     KeyPair server = KeyPair.random();
     KeyPair client = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -733,7 +766,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(client.getAccountId())
         .build();
 
@@ -751,10 +784,43 @@ public class Sep10ChallengeTest {
     String challenge = transaction.toEnvelopeXdrBase64();
 
     try {
-      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET);
+      Sep10Challenge.readChallengeTransaction(challenge, server.getAccountId(), Network.TESTNET, domainName);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Random nonce before encoding as base64 should be 48 bytes long.", e.getMessage());
+    }
+  }
+
+  @Test
+  public void testReadChallengeTransactionInvalidDomainNameMismatch() throws IOException {
+    KeyPair server = KeyPair.random();
+    KeyPair client = KeyPair.random();
+    Network network = Network.TESTNET;
+    String domainName = "example.com";
+    String mismatchDomainName = "mismatch_example.com";
+
+    long now = System.currentTimeMillis() / 1000L;
+    long end = now + 300;
+    TimeBounds timeBounds = new TimeBounds(now, end);
+
+    Transaction transaction = null;
+    try {
+      transaction = Sep10Challenge.newChallenge(
+          server,
+          network,
+          client.getAccountId(),
+          domainName,
+          timeBounds
+      );
+    } catch (InvalidSep10ChallengeException e) {
+      fail("Should not have thrown any exception.");
+    }
+
+    try {
+      Sep10Challenge.readChallengeTransaction(transaction.toEnvelopeXdrBase64(), server.getAccountId(), Network.TESTNET, mismatchDomainName);
+      fail();
+    } catch (InvalidSep10ChallengeException e) {
+      assertEquals("The transaction's operation key name does not include the expected home domain.", e.getMessage());
     }
   }
 
@@ -765,7 +831,7 @@ public class Sep10ChallengeTest {
     KeyPair masterClient = KeyPair.random();
     KeyPair signerClient1 = KeyPair.random();
     KeyPair signerClient2 = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
 
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
@@ -778,7 +844,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(masterClient.getAccountId())
         .build();
 
@@ -803,7 +869,7 @@ public class Sep10ChallengeTest {
 
     int threshold = 6;
     try {
-      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals(String.format("Transaction not signed by server: %s.", server.getAccountId()), e.getMessage());
@@ -820,12 +886,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -836,7 +903,7 @@ public class Sep10ChallengeTest {
     ));
 
     int threshold = 255;
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
     assertEquals(new HashSet<String>(Collections.singletonList(masterClient.getAccountId())), signersFound);
   }
 
@@ -850,12 +917,14 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
+
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -870,7 +939,7 @@ public class Sep10ChallengeTest {
     ));
 
     int threshold = 7;
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
     assertEquals(new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId())), signersFound);
   }
 
@@ -884,12 +953,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -903,7 +973,7 @@ public class Sep10ChallengeTest {
     ));
 
     int threshold = 3;
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
     assertEquals(new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId())), signersFound);
   }
 
@@ -918,12 +988,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -943,7 +1014,7 @@ public class Sep10ChallengeTest {
     ));
 
     int threshold = 3;
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
     assertEquals(new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId())), signersFound);
   }
 
@@ -958,12 +1029,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -978,7 +1050,7 @@ public class Sep10ChallengeTest {
 
     int threshold = 7;
     try {
-      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Signers with weight 3 do not meet threshold 7.", e.getMessage());
@@ -996,12 +1068,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1018,7 +1091,7 @@ public class Sep10ChallengeTest {
 
     int threshold = 7;
     try {
-      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction has unrecognized signatures.", e.getMessage());
@@ -1036,12 +1109,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1052,7 +1126,7 @@ public class Sep10ChallengeTest {
     Set<Sep10Challenge.Signer> signers = Collections.emptySet();
     int threshold = 3;
     try {
-      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("No verifiable signers provided, at least one G... address must be provided.", e.getMessage());
@@ -1070,12 +1144,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1094,7 +1169,7 @@ public class Sep10ChallengeTest {
     int threshold = 3;
 
     try {
-      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+      Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("No verifiable signers provided, at least one G... address must be provided.", e.getMessage());
@@ -1110,12 +1185,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1128,7 +1204,7 @@ public class Sep10ChallengeTest {
     ));
 
     int threshold = 1;
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, threshold, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionThreshold(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, threshold, signers);
     assertEquals(new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId())), signersFound);
   }
 
@@ -1138,7 +1214,8 @@ public class Sep10ChallengeTest {
     KeyPair masterClient = KeyPair.random();
     KeyPair signerClient1 = KeyPair.random();
     KeyPair signerClient2 = KeyPair.random();
-    String anchorName = "Stellar Test";
+    String domainName = "example.com";
+
     Network network = Network.TESTNET;
 
     long now = System.currentTimeMillis() / 1000L;
@@ -1152,7 +1229,7 @@ public class Sep10ChallengeTest {
     byte[] encodedNonce = base64Encoding.encode(nonce).getBytes();
 
     Account sourceAccount = new Account(server.getAccountId(), -1L);
-    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(anchorName + " auth", encodedNonce)
+    ManageDataOperation manageDataOperation1 = new ManageDataOperation.Builder(domainName + " auth", encodedNonce)
         .setSourceAccount(masterClient.getAccountId())
         .build();
 
@@ -1173,7 +1250,7 @@ public class Sep10ChallengeTest {
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId(), KeyPair.random().getAccountId()));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals(String.format("Transaction not signed by server: %s.", server.getAccountId()), e.getMessage());
@@ -1189,19 +1266,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(masterClient);
 
     Set<String> signers = new HashSet<String>(Collections.singletonList(masterClient.getAccountId()));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(signers, signersFound);
   }
 
@@ -1216,18 +1294,19 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId(), KeyPair.random().getAccountId()));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction not signed by any client signer.", e.getMessage());
@@ -1245,12 +1324,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1261,7 +1341,7 @@ public class Sep10ChallengeTest {
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId(), KeyPair.random().getAccountId()));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction has unrecognized signatures.", e.getMessage());
@@ -1279,12 +1359,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1292,7 +1373,7 @@ public class Sep10ChallengeTest {
     transaction.sign(signerClient2);
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId(), KeyPair.random().getAccountId()));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(new HashSet<String>(Arrays.asList(signerClient1.getAccountId(), signerClient2.getAccountId())), signersFound);
   }
 
@@ -1307,12 +1388,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1320,7 +1402,7 @@ public class Sep10ChallengeTest {
     transaction.sign(signerClient1);
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), signerClient2.getAccountId(), KeyPair.random().getAccountId()));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(new HashSet<String>(Arrays.asList(signerClient1.getAccountId(), signerClient2.getAccountId())), signersFound);
   }
 
@@ -1335,19 +1417,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(signerClient1);
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId()));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(new HashSet<String>(Collections.singletonList(signerClient1.getAccountId())), signersFound);
   }
 
@@ -1361,19 +1444,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(signerClient1);
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), server.getAccountId()));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(new HashSet<String>(Collections.singletonList(signerClient1.getAccountId())), signersFound);
   }
 
@@ -1388,18 +1472,19 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId(), server.getAccountId()));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction not signed by any client signer.", e.getMessage());
@@ -1417,12 +1502,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1433,7 +1519,7 @@ public class Sep10ChallengeTest {
     String unknownSignerType = "?ARPF6NZRR7EEVO7ESIWUDXHAOMM2QSKIQQBJK6I2FB7YKDZES5UCLWD";
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(),
         signerClient1.getAccountId(), preauthTxHash, xHash, unknownSignerType));
-    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+    Set<String> signersFound = Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
     assertEquals(new HashSet<String>(Collections.singletonList(signerClient1.getAccountId())), signersFound);
   }
 
@@ -1447,12 +1533,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1461,7 +1548,7 @@ public class Sep10ChallengeTest {
 
     Set<String> signers = new HashSet<String>(Arrays.asList(masterClient.getAccountId(), signerClient1.getAccountId()));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("Transaction has unrecognized signatures.", e.getMessage());
@@ -1478,12 +1565,13 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
@@ -1491,7 +1579,7 @@ public class Sep10ChallengeTest {
 
     Set<String> signers = new HashSet<String>(Collections.singletonList(new String(signerClient1.getSecretSeed())));
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, signers);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, signers);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("No verifiable signers provided, at least one G... address must be provided.", e.getMessage());
@@ -1504,6 +1592,7 @@ public class Sep10ChallengeTest {
     KeyPair masterClient = KeyPair.random();
     KeyPair signerClient1 = KeyPair.random();
     Network network = Network.TESTNET;
+    String domainName = "example.com";
 
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
@@ -1513,14 +1602,14 @@ public class Sep10ChallengeTest {
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(signerClient1);
 
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, null);
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, null);
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("No verifiable signers provided, at least one G... address must be provided.", e.getMessage());
@@ -1537,19 +1626,20 @@ public class Sep10ChallengeTest {
     long now = System.currentTimeMillis() / 1000L;
     long end = now + 300;
     TimeBounds timeBounds = new TimeBounds(now, end);
+    String domainName = "example.com";
 
     Transaction transaction = Sep10Challenge.newChallenge(
         server,
         network,
         masterClient.getAccountId(),
-        "Stellar Test",
+        domainName,
         timeBounds
     );
 
     transaction.sign(signerClient1);
 
     try {
-      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, Collections.<String>emptySet());
+      Sep10Challenge.verifyChallengeTransactionSigners(transaction.toEnvelopeXdrBase64(), server.getAccountId(), network, domainName, Collections.<String>emptySet());
       fail();
     } catch (InvalidSep10ChallengeException e) {
       assertEquals("No verifiable signers provided, at least one G... address must be provided.", e.getMessage());